filtered_string: Normalized version of the text
This is a copy of the string with only the characters that belong to the
given charset. This allow to work faster on it.

With it, a function that replace all the relevant characters with a new
normalized text. i.e. give the normalized text, this function insert all
the decorative characters.

Signed-off-by: Celelibi <celelibi@gmail.com>